moar cleanup...

+++ vecsum/parallel-mpi.c
@@ -17,8 +17,7 @@ int isPowerOfTwo (unsigned int x)
   return ((x != 0) && !(x & (x - 1)));
 }
 
-int main(int argc, char** argv)
-{
+int main(int argc, char** argv){
     int rank = 0,size = 1;
     double pi = 4.0*atan(1), sum = pi*pi/6;
     double time_init;
@@ -26,24 +25,24 @@ int main(int argc, char** argv)
     MPI_Comm_size(MPI_COMM_WORLD, &size);
     MPI_Comm_rank(MPI_COMM_WORLD, &rank);
 
-    if (rank == 0){
+    if(rank == 0){
         printf("MPI   \tThreadcount: %i\n",size);
-        if (argc < 2) {
+        if(argc < 2) {
             printf("Need one parameter, the size of the vector\n");
             MPI_Finalize();
             return 1;
-        }else if (!isPowerOfTwo(size)){
+        }else if(!isPowerOfTwo(size)){
             printf("The number of processors must be a power of 2, %i processors supplied.\n",size);
             MPI_Finalize();
             return 1;
         }
     }
-    long int N = atoi(argv[1]), share = N/size;
+    int N = atoi(argv[1]), share = N/size;
     double sumn = 0.0;
     double* v = (double*)calloc(share,sizeof(double));
     time_init = walltime();
 
-    for (long int i=share; i>0; i--){
+    for(int i=share; i>0; i--){
         v[i-1] = 1.0/(((double)i+rank*share)*(i+rank*share));
         sumn += v[i-1];
     }
@@ -54,7 +53,7 @@ int main(int argc, char** argv)
         printf("Error:\t\t%e \nTime Elapsed:\t%f \n",sum-sumn,walltime()-time_init);
     }
     MPI_Finalize();
-return 0;
+    return 0;
 }
