+++ vecsum/serial.c
@@ -9,17 +9,16 @@ double walltime ( void ) {  // Returns current time, with microsecond precision
     return ( t.tv_sec + 1e-6 * t.tv_usec );
 }
 
-int main(int argc, char** argv)
-{
+int main(int argc, char** argv){
     double pi = 4.0*atan(1);
     double sum = pi*pi/6;
     double time_init;
 
-    if (argc < 2) {
+    if(argc < 2) {
         printf("Need one parameter, the size of the vector\n");
         return 1;
     }
-    long int n = atoi(argv[1]);
+    int n = atoi(argv[1]);
     printf("Serial:\n");
     double* v = (double*)malloc(n*sizeof(double));
     double sumn = 0;
@@ -30,5 +29,5 @@ int main(int argc, char** argv)
         sumn += v[i];
     }
     printf("Error:\t\t%e \nTime Elapsed:\t%f\n",sum-sumn,walltime()-time_init);
-return 0;
+    return 0;
 }
